requirements.txt
strands-agents
bedrock-agentcore
mcp
boto3
python-dotenv